test: completed testing for separated types and refactored the test utils for sharing across the crate.
