+++ src/calculator/mod.rs
@@ -18,7 +18,7 @@ mod route;
 
 pub use self::force_level::ForceLevel;
 use self::route::CalcRoute;
-use self::{conventional::ConventionalCommits, next_version::NextVersion};
+pub(crate) use self::{conventional::ConventionalCommits, next_version::NextVersion};
 pub use level_hierarchy::LevelHierarchy;
 
 use crate::{version::PreReleaseType, Error, Level, VersionTag};
@@ -547,178 +547,16 @@ impl VersionCalculator {
 
 #[cfg(test)]
 mod test {
-    use std::collections::{HashMap, HashSet};
-    use std::ffi::OsString;
-    use std::fmt;
-    use std::str::FromStr;
-
     use log::LevelFilter;
     use log4rs_test_utils::test_logging;
     use rstest::rstest;
 
+    use super::VersionCalculator;
     use crate::calculator::CalcRoute;
+    use crate::test_utils;
+    use crate::test_utils::*;
     use crate::version::PreRelease;
-    use crate::LevelHierarchy::Feature;
-    use crate::{version::Semantic, VersionCalculator, VersionTag};
-    use crate::{ForceLevel, LevelHierarchy};
-
-    use super::conventional::ConventionalCommits;
-
-    #[derive(Debug)]
-    pub(crate) enum ConventionalType {
-        Feat,
-        Fix,
-        Docs,
-        Style,
-        Refactor,
-        Perf,
-        Test,
-        Build,
-        Ci,
-        Chore,
-        Revert,
-    }
-
-    impl fmt::Display for ConventionalType {
-        fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-            match self {
-                ConventionalType::Feat => write!(f, "feat"),
-                ConventionalType::Fix => write!(f, "fix"),
-                ConventionalType::Docs => write!(f, "docs"),
-                ConventionalType::Style => write!(f, "style"),
-                ConventionalType::Refactor => write!(f, "refactor"),
-                ConventionalType::Perf => write!(f, "perf"),
-                ConventionalType::Test => write!(f, "test"),
-                ConventionalType::Build => write!(f, "build"),
-                ConventionalType::Chore => write!(f, "chore"),
-                ConventionalType::Ci => write!(f, "ci"),
-                ConventionalType::Revert => write!(f, "revert"),
-            }
-        }
-    }
-
-    // Implement the FromStr trait for the Direction enum
-    impl FromStr for ConventionalType {
-        type Err = ();
-
-        fn from_str(s: &str) -> Result<Self, Self::Err> {
-            match s.trim().to_lowercase().as_str() {
-                "feat" => Ok(ConventionalType::Feat),
-                "fix" => Ok(ConventionalType::Fix),
-                "docs" => Ok(ConventionalType::Docs),
-                "style" => Ok(ConventionalType::Style),
-                "refactor" => Ok(ConventionalType::Refactor),
-                "perf" => Ok(ConventionalType::Perf),
-                "test" => Ok(ConventionalType::Test),
-                "build" => Ok(ConventionalType::Build),
-                "chore" => Ok(ConventionalType::Chore),
-                "ci" => Ok(ConventionalType::Ci),
-                "revert" => Ok(ConventionalType::Revert),
-                _ => Err(()),
-            }
-        }
-    }
-
-    fn gen_current_version(
-        version_prefix: &str,
-        major: u32,
-        minor: u32,
-        patch: u32,
-        pre_release: Option<PreRelease>,
-        build_meta_data: Option<String>,
-    ) -> VersionTag {
-        VersionTag {
-            refs: "refs/tags/".to_string(),
-            tag_prefix: "".to_string(),
-            version_prefix: version_prefix.to_string(),
-            semantic_version: Semantic {
-                major,
-                minor,
-                patch,
-                pre_release,
-                build_meta_data,
-            },
-        }
-    }
-
-    fn gen_conventional_commits() -> Option<ConventionalCommits> {
-        let mut counts = HashMap::new();
-        let values = [
-            ("chore".to_string(), 1),
-            ("docs".to_string(), 1),
-            ("feat".to_string(), 1),
-            ("refactor".to_string(), 1),
-            ("ci".to_string(), 1),
-            ("test".to_string(), 1),
-            ("fix".to_string(), 1),
-        ];
-
-        for val in values {
-            counts.insert(val.0, val.1);
-        }
-
-        Some(
-            ConventionalCommits {
-                commits: vec!(
-                    "fix: spelling of output in description of set_env".to_string(),
-                    "Merge branch 'main' of github.com:jerusdp/nextsv into fix/version-level-assessment".to_string(),
-                    "test: Ensure all current tests are passing".to_string(),
-                    "refactor: implemented VersionTag".to_string(),
-                    "feat: Regex implemented to extract version string".to_string(),
-                    "chore: Updated minium rust version references".to_string(),
-                    "ci: Updated Minimum rust version to 1.74".to_string(),
-                    "docs: Updated tests in docs.".to_string()
-                ),
-                counts,
-                breaking: false,
-                top_type: Some(
-                    Feature,
-                ),
-            },
-        )
-    }
-
-    fn gen_conventional_commit(
-        commit_type: ConventionalType,
-        breaking: bool,
-    ) -> Option<ConventionalCommits> {
-        let mut counts = HashMap::new();
-        counts.insert(format!("{}", commit_type), 1);
-
-        let commits = vec![format!(
-            "{}{} commit for testing purposes only",
-            if breaking { "!" } else { "" },
-            breaking
-        )];
-
-        let top_type = Some(LevelHierarchy::parse(&commit_type.to_string()).unwrap());
-
-        Some(ConventionalCommits {
-            commits,
-            counts,
-            breaking,
-            top_type,
-        })
-    }
-
-    fn gen_files() -> Option<HashSet<OsString>> {
-        let file_list = [
-            "calculator.rs",
-            "help.trycmd",
-            "CHANGELOG.md",
-            "Cargo.toml",
-            "config.yml",
-            "error.rs",
-            "README.md",
-        ];
-        let mut files = HashSet::new();
-
-        for file in file_list {
-            files.insert(OsString::from(file));
-        }
-
-        Some(files)
-    }
+    use crate::ForceLevel;
 
     #[rstest]
     #[case::feature("feat", "minor", "0.8.0")]
@@ -737,9 +575,9 @@ mod test {
         #[case] expected_level: &str,
         #[case] expected_version: &str,
     ) {
-        let current_version = gen_current_version("v", 0, 7, 9, None, None);
-        let conventional = gen_conventional_commit(commit, false);
-        let files = gen_files();
+        let current_version = test_utils::gen_current_version("v", 0, 7, 9, None, None);
+        let conventional = test_utils::gen_conventional_commit(commit, false);
+        let files = test_utils::gen_files();
 
         let mut calculator = VersionCalculator {
             current_version,
@@ -769,9 +607,9 @@ mod test {
     #[case::revert("revert")]
     // #[trace]
     fn bump_result_for_nonprod_current_version_and_breaking(#[case] commit: ConventionalType) {
-        let current_version = gen_current_version("v", 0, 7, 9, None, None);
-        let conventional = gen_conventional_commit(commit, true);
-        let files = gen_files();
+        let current_version = test_utils::gen_current_version("v", 0, 7, 9, None, None);
+        let conventional = test_utils::gen_conventional_commit(commit, true);
+        let files = test_utils::gen_files();
 
         let mut calculator = VersionCalculator {
             current_version,
@@ -804,9 +642,9 @@ mod test {
         #[case] expected_bump: &str,
         #[case] expected_version: &str,
     ) {
-        let current_version = gen_current_version("v", 1, 7, 9, None, None);
-        let conventional = gen_conventional_commit(commit, false);
-        let files = gen_files();
+        let current_version = test_utils::gen_current_version("v", 1, 7, 9, None, None);
+        let conventional = test_utils::gen_conventional_commit(commit, false);
+        let files = test_utils::gen_files();
 
         let mut calculator = VersionCalculator {
             current_version,
@@ -833,9 +671,9 @@ mod test {
     ) {
         test_logging::init_logging_once_for(vec![], LevelFilter::Debug, None);
         let current_version =
-            gen_current_version("v", 0, 7, 9, Some(PreRelease::new("alpha.1")), None);
-        let conventional = gen_conventional_commit(commit, false);
-        let files = gen_files();
+            test_utils::gen_current_version("v", 0, 7, 9, Some(PreRelease::new("alpha.1")), None);
+        let conventional = test_utils::gen_conventional_commit(commit, false);
+        let files = test_utils::gen_files();
 
         let route = CalcRoute::new(&current_version.semantic_version);
 
@@ -865,9 +703,9 @@ mod test {
         commit: ConventionalType,
     ) {
         test_logging::init_logging_once_for(vec![], LevelFilter::Debug, None);
-        let current_version = gen_current_version("v", 1, 7, 9, None, None);
-        let conventional = gen_conventional_commit(commit, true);
-        let files = gen_files();
+        let current_version = test_utils::gen_current_version("v", 1, 7, 9, None, None);
+        let conventional = test_utils::gen_conventional_commit(commit, true);
+        let files = test_utils::gen_files();
 
         let mut calculator = VersionCalculator {
             current_version,
@@ -886,9 +724,9 @@ mod test {
     #[test]
     fn promote_to_version_one() {
         test_logging::init_logging_once_for(vec![], LevelFilter::Debug, None);
-        let current_version = gen_current_version("v", 0, 7, 9, None, None);
-        let conventional = gen_conventional_commits();
-        let files = gen_files();
+        let current_version = test_utils::gen_current_version("v", 0, 7, 9, None, None);
+        let conventional = test_utils::gen_conventional_commits();
+        let files = test_utils::gen_files();
 
         let mut calculator = VersionCalculator {
             current_version,
